Add files via upload
Hi R Ladies Indonesia!!
Welcome to the club!!
It's really great that after 7 months running, the R Ladies Indonesia have developed two separate chapters (not too separated, but still ;)) in Bogor and in Jakarta. We are happy to meet new faces during our meetups, and we'd like to support all the newcomers while moving forwards first with our current regular participants. 

As you might have noticed, our R Ladies, be it the one in Bogor or Jakarta, has a running theme of intro to R - classical statistics with R - advance statistics with R before we decide on what's next to have in our meetups (maybe we'll dwell into RMarkdown and Tidyverse). Since we need to keep on moving, we are unfortunately aren't capable to give intro to R at every meetup. 

Therefore, we would like to ask you, interested participants, to help us by first study the intro to R materials, and run some very basic codes to help you get started and help you to catch up with the ongoing themes. 

To do so, we have prepared this github repository, and what you need to do is:
1. download all the files in this repository
2. read through "Intro to R.pdf"
3. read through "Starter Kit R.pdf" and do the task at the last page

This is to ensure that you're set to join the meetup and also help the regular participants to focus on the planned themes. 

We will most likely held an absolute beginner to R meetup next year where you can refresh all the basic R knowledge, but until then, it'll be great if you do step 1, 2, 3 above. 

Many thanks, and looking forward to see you at our meetup!!

Best,
Ulfah Mardhiah
R Ladies Bogor & R Ladies Jakarta coordinator